pyproject.toml
[project]
name = "transcript-combobulator"
version = "0.4.1"
description = "Pre-processes audio files for better Whisper transcription"
readme = "docs/README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = []
authors = [
    { name = "Sean Kennedy", email = "sean@codeserious.com" }
]
dependencies = [
    "openai-whisper==20250625",
    "python-dotenv==1.2.1",
    "soundfile==0.13.1",
    "torch==2.10.0",
    "torchaudio==2.10.0",
    "webvtt-py==0.5.1",
    "silero-vad==6.2.0",
    "torchcodec>=0.10.0",
    "rich>=13.0",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]

[project.optional-dependencies]
dev = [
    "mypy==1.19.1",
    "pytest==9.0.2",
]

[project.urls]
Homepage = "https://github.com/seriouslysean/transcript-combobulator"
Repository = "https://github.com/seriouslysean/transcript-combobulator.git"
Issues = "https://github.com/seriouslysean/transcript-combobulator/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"
pythonpath = ["."]

[tool.mypy]
python_version = "3.10"
# Annotation coverage is not enforced — mypy runs advisory and catches obvious
# bugs (undefined names, return type mismatches on annotated funcs) without
# blocking on every untyped function.
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_no_return = true
warn_unreachable = true
ignore_missing_imports = true

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
multi_line_output = 3
